Add 'get' Method to the RailsResource Prototype
This is a total stab in the dark for adding a get method to
resource instances so that server-side changes can be requested
and applied to an instance of a resource.

I'm hoping this might help attract someone who is familiar with
the project to take a look at Issue #112. This might fix issue #112
but, again, this is just a stab in the dark.